# libs/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Trader API",
    "summary": "Исследовательская платформа по рыночным данным",
    "description": "REST API платформы. Документация: `/docs` (Swagger UI), `/redoc`, схема — `/openapi.json`.\n\n**WebSocket** `/ws/jobs/{job_id}` (в OpenAPI не описывается): шлёт JSON в формате `JobOut` при каждом изменении задачи и закрывается по её завершении; код закрытия 4404 — задачи нет.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/indicator-values": {
      "get": {
        "tags": [
          "indicators"
        ],
        "summary": "Значения индикатора на баре графика",
        "description": "Считает индикатор по барам `source_timeframe` (по умолчанию как у графика) и проецирует на последние `limit` баров `chart_timeframe` ступенькой по `available_at` (без формирующегося старшего бара). С `as_of` — только закрытые к моменту бары. Значения до конца прогрева — `valid = false`.",
        "operationId": "getIndicatorValues",
        "parameters": [
          {
            "name": "indicator",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Имя из `GET /indicators`",
              "title": "Indicator"
            },
            "description": "Имя из `GET /indicators`"
          },
          {
            "name": "chart_timeframe",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "15m, 1h, 4h, 1d, 1w",
              "title": "Chart Timeframe"
            },
            "description": "15m, 1h, 4h, 1d, 1w"
          },
          {
            "name": "instrument_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Instrument Id"
            }
          },
          {
            "name": "source_timeframe",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "По умолчанию равен chart_timeframe",
              "title": "Source Timeframe"
            },
            "description": "По умолчанию равен chart_timeframe"
          },
          {
            "name": "params",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Параметры индикатора JSON-объектом",
              "default": "{}",
              "title": "Params"
            },
            "description": "Параметры индикатора JSON-объектом"
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20000,
              "minimum": 1,
              "default": 2000,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndicatorValuesOut"
                }
              }
            }
          },
          "404": {
            "description": "Не найдено"
          },
          "422": {
            "description": "Неверные параметры индикатора, инструмент или таймфреймы (source TF младше chart TF запрещён)"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IndicatorPoint": {
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "Open time бара графика"
          },
          "values": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Values"
          },
          "valid": {
            "type": "boolean",
            "title": "Valid",
            "description": "Прогрев source-серии завершён"
          },
          "source_timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Timestamp",
            "description": "Open time source-бара, давшего значение"
          },
          "available_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available At",
            "description": "Когда значение стало известно (закрытие source-бара)"
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "values",
          "valid",
          "source_timestamp",
          "available_at"
        ],
        "title": "IndicatorPoint"
      },
      "IndicatorValuesOut": {
        "properties": {
          "indicator": {
            "type": "string",
            "title": "Indicator"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          },
          "params_hash": {
            "type": "string",
            "title": "Params Hash"
          },
          "pane": {
            "type": "string",
            "title": "Pane"
          },
          "outputs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Outputs"
          },
          "chart_timeframe": {
            "type": "string",
            "title": "Chart Timeframe"
          },
          "source_timeframe": {
            "type": "string",
            "title": "Source Timeframe"
          },
          "warmup_bars": {
            "type": "integer",
            "title": "Warmup Bars"
          },
          "source_bar_count": {
            "type": "integer",
            "title": "Source Bar Count",
            "description": "Баров source TF использовано в расчёте"
          },
          "source_truncated": {
            "type": "boolean",
            "title": "Source Truncated",
            "description": "История source TF длиннее лимита: прогрев начат не с начала данных"
          },
          "as_of": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/IndicatorPoint"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "indicator",
          "version",
          "params",
          "params_hash",
          "pane",
          "outputs",
          "chart_timeframe",
          "source_timeframe",
          "warmup_bars",
          "source_bar_count",
          "source_truncated",
          "as_of",
          "points"
        ],
        "title": "IndicatorValuesOut"
      },
...
    }
  },
...
}